Refactor RSS feed elements to separate file.
Our django.contrib.syndication.views.Feed subclass was not unit testable as it was defined within a class method.
